splitted test files and added new tests and implementation of MongoDatabaseExtensions CollectionExists method
